1214 검색 기능 추가
