+++ src/components/Header/TopHeader.jsx
@@ -1,8 +1,9 @@
 import styles from "./TopHeader.module.css";
 import Diversity2OutlinedIcon from "@mui/icons-material/Diversity2Outlined";
 import { Link } from "react-router-dom";
+import SearchInput from "../modal/SearchInput";
+//모바일 환경에서의 헤더 + 검색 기능
 const TopHeader = () => {
-  //모바일 환경에서의 헤더 + 검색 기능
   return (
     <header className={styles.container}>
       <Link to="/">
@@ -11,10 +12,9 @@ const TopHeader = () => {
           <h1>travel</h1>
         </div>
       </Link>
-      {/* 헤더 */}
-
-      <input type="text" className={styles.search_input} />
-      {/* 검색 기능 추가 필요 */}
+      <div className={styles.search_wrap}>
+        <SearchInput />
+      </div>
     </header>
   );
 };
