+++ src/components/Header/TopHeader.module.css
@@ -11,17 +11,10 @@
   font-size: var(--font-xxlarge);
   margin: 0 10px;
 }
-.search_input {
-  height: 35px;
-  width: 40%;
-  margin-right: 20px;
-  border: 1px solid var(--color-gray);
-  border-radius: 5px;
-  padding: 5px 10px;
+.search_wrap{
+  width : 300px;
 }
-.search_input:focus {
-  outline: 1px solid var(--color-gray);
-}
+
 @media (max-width: 768px) {
   .container {
     display: flex;
